added checking of existence of temporary files; improved logging; code cleanup;

+++ scripts/build-lm.sh
@@ -1,6 +1,7 @@
 #! /bin/bash
 
 set -m # Enable Job Control
+set -e
 
 function usage()
 {
@@ -29,10 +30,31 @@ OPTIONS:
        --no-zipping            disabling zipping of files (default: enabled)
        --add-start-end         add start and end symbols before and after each line
        -v|--verbose            Verbose
+       --debug                 Enable debug (deafult false); if enabled the temporary files are not removed
        -irstlm|--irstlm        optionally set the directory of installation of IRSTLM; if not specified, the environment variable IRSTLM is used
        -h|-?|--help            Show this message
 
 EOF
+}
+
+function fileExists()
+{
+    file=$1; shift;
+
+    if [ ! -e $file ] ; then
+        echo "file ($file) does not exist" >> $logfile 2>&1
+        exit 1
+    else
+        if [ $debug ] ; then
+            echo "file ($file) exists" >> $logfile 2>&1
+        fi
+    fi
+}
+
+function fileIsOpened()
+{
+    file=$1; shift;
+    $IRSTLM/bin/check_if_closed.sh $file
 }
 
 #default parameters
@@ -44,6 +66,7 @@ inpfile="";
 outfile=""
 zipping="--zipping";
 addstartend="";
+debug="";
 verbose="";
 smoothing="witten-bell";
 prune="";
@@ -88,16 +111,18 @@ while [ "$1" != "" ]; do
                                 ;;
         -b | --boundaries )     boundaries='--cross-sentence';
 				;;
-        --zipping )        zipping='--zipping';
-				;;
-        --no-zipping )        zipping='';
-				;;
+        --zipping )         zipping='--zipping';
+				            ;;
+        --no-zipping )      zipping='';
+	                        ;;
         --add-start-end )       addstartend='--addstartend';
-				;;
+				                ;;
         -v | --verbose )        verbose='--verbose';
                                 ;;
+        --debug )               debug='--debug';
+				                ;;
         -irstlm | --irstlm )	shift;
-				IRSTLM=$1;
+				                IRSTLM=$1;
                                 ;;
         -h | -? | --help )      usage;
                                 exit 0;
@@ -107,6 +132,17 @@ while [ "$1" != "" ]; do
     esac
     shift
 done
+
+if [ -e $logfile -a $logfile != "/dev/null"  -a $logfile != "/dev/stderr" -a $logfile != "/dev/stdout" ]; then
+   echo "Logfile $logfile already exists! either remove or rename it."
+   exit 7
+fi
+
+if [ $debug ] ; then
+    if [ $logfile == "/dev/null" ] ; then
+        logfile="/dev/stderr"
+    fi
+fi
 
 if [ ! $IRSTLM ]; then
    echo "Set IRSTLM environment variable with path to irstlm"
@@ -154,13 +190,12 @@ exit 4
 esac
 			
 
-echo "LOGFILE:$logfile"
-if [ $zipping ] ; then
-echo "ZIPPING:yes"
-else
-echo "ZIPPING:no"
-fi		 
 
+if [ $zipping ] ; then
+echo "ZIPPING:yes" >> $logfile 2>&1
+else
+echo "ZIPPING:no" >> $logfile 2>&1
+fi
 
 if [ ! "$inpfile" -o ! "$outfile" ] ; then
     usage
@@ -172,10 +207,6 @@ if [ -e $outfile ]; then
    exit 6
 fi
 
-if [ -e $logfile -a $logfile != "/dev/null" -a $logfile != "/dev/stdout" ]; then
-   echo "Logfile $logfile already exists! either remove or rename it."
-   exit 7
-fi
 
 if [ $verbose ] ; then
 echo inpfile='"'$inpfile'"' outfile=$outfile order=$order parts=$parts tmpdir=$tmpdir prune=$prune smoothing=$smoothing dictionary=$dictionary verbose=$verbose prune_thr_str=$prune_thr_str zipping=$zipping addstartend=$addstartend >> $logfile 2>&1
@@ -199,38 +230,54 @@ fi
 res=`echo $inpfile | grep ' '`
 actual_inpfile=""
 if [[ $res ]] ; then
-echo "the inpfile is a command, actually"
-if [ $addstartend ] ; then
-actual_inpfile="$inpfile | $scr/add-start-end.sh"
+    echo "the inpfile is a command, actually" >> $logfile 2>&1
+    if [ $addstartend ] ; then
+        actual_inpfile="$inpfile | $scr/add-start-end.sh"
+    else
+        actual_inpfile="$inpfile"
+    fi
 else
-actual_inpfile="$inpfile"
-fi
-else
-echo "the inpfile is a file, actually"
-if [ $addstartend ] ; then
-actual_inpfile="$scr/add-start-end.sh < $inpfile"
-else
-actual_inpfile="$inpfile"
+    echo "the inpfile is a file, actually" >> $logfile 2>&1
+    if [ $addstartend ] ; then
+        actual_inpfile="$scr/add-start-end.sh < $inpfile"
+    else
+        actual_inpfile="$inpfile"
+    fi
 fi
 
-fi
 
-echo "input is the following:\'$actual_inpfile\'"
+if [ $verbose ] ; then
+    echo "input is the following:\'$actual_inpfile\'" >> $logfile 2>&1
+fi
 
 date >> $logfile 2>&1
 echo "Extracting dictionary from training corpus" >> $logfile 2>&1
 $bin/dict -i="$actual_inpfile" -o=$tmpdir/dictionary $uniform -sort=no 2> $logfile
 
+res=$?
+if [ $res -ne 0 ] ; then
+    echo "creation of the dictionary failed with error status $res"
+    exit $res
+fi
+
+
 date >> $logfile 2>&1
 echo "Splitting dictionary into $parts lists" >> $logfile 2>&1
-$scr/split-dict.pl --input $tmpdir/dictionary --output $tmpdir/dict. --parts $parts >> $logfile 2>&1
+sfxList=`$scr/split-dict.pl --input $tmpdir/dictionary --output $tmpdir/dict. --parts $parts 2>> $logfile`
+echo "sfxList:$sfxList" >> $logfile 2>&1
+
+for sfx in $sfxList ; do
+    fileExists $tmpdir/dict.${sfx}
+    fileIsOpened $tmpdir/dict.${sfx}
+done
 
 date >> $logfile 2>&1
 echo "Extracting n-gram statistics for each word list" >> $logfile 2>&1
 echo "Important: dictionary must be ordered according to order of appearance of words in data" >> $logfile 2>&1
 echo "used to generate n-gram blocks,  so that sub language model blocks results ordered too" >> $logfile 2>&1
 
-for sdict in $tmpdir/dict.*;do
+for sfx in $sfxList ; do
+sdict="$tmpdir/dict.$sfx"
 sdict=`basename $sdict`
 echo "Extracting n-gram statistics for $sdict" >> $logfile 2>&1
 if [ $smoothing = "--shift-beta" -o $smoothing = "--improved-shift-beta" ]; then
@@ -248,6 +295,16 @@ done
 
 # Wait for all parallel jobs to finish
 while [ 1 ]; do fg 2> /dev/null; [ $? == 1 ] && break; done
+
+for sfx in $sfxList ; do
+    if [ $zipping ] ; then
+        fileExists $tmpdir/ngram.dict.${sfx}.gz
+        fileIsOpened $tmpdir/ngram.dict.${sfx}.gz
+    else
+        fileExists $tmpdir/ngram.dict.${sfx}
+        fileIsOpened $tmpdir/ngram.dict.${sfx}
+    fi
+done
 
 date >> $logfile 2>&1
 echo "Estimating language models for each word list" >> $logfile 2>&1
@@ -274,6 +331,22 @@ done
 # Wait for all parallel jobs to finish
 while [ 1 ]; do fg 2> /dev/null; [ $? == 1 ] && break; done
 
+
+# check whether all files for the sublm are created
+for sfx in $sfxList ; do
+    o=1
+    while [ $o -le $order ] ; do
+        if [ $zipping ] ; then
+            fileExists $tmpdir/lm.dict.${sfx}.${o}gr.gz
+            fileIsOpened $tmpdir/lm.dict.${sfx}.${o}gr.gz
+        else
+            fileExists $tmpdir/lm.dict.${sfx}.${o}gr
+            fileIsOpened $tmpdir/lm.dict.${sfx}.${o}gr
+        fi
+        o=$(( $o + 1))
+    done
+done
+
 date >> $logfile 2>&1
 echo "Merging language models into $outfile" >> $logfile 2>&1
 if [ $zipping ] ; then
@@ -282,15 +355,25 @@ else
 $scr/merge-sublm.pl --size $order --sublm $tmpdir/lm.dict -lm $outfile $backoff  >> $logfile 2>&1
 fi
 
+if [ $zipping ] ; then
+    fileExists $outfile.gz
+    fileIsOpened $outfile.gz
+else
+    fileExists $outfile
+    fileIsOpened $outfile
+fi
 date >> $logfile 2>&1
-echo "Cleaning temporary directory $tmpdir" >> $logfile 2>&1
-rm $tmpdir/* 2> /dev/null
 
-if [ $tmpdir_created -eq 1 ]; then
-    echo "Removing temporary directory $tmpdir" >> $logfile 2>&1
-    rmdir $tmpdir 2> /dev/null
-    if [ $? != 0 ]; then
-        echo "Warning: the temporary directory could not be removed." >> $logfile 2>&1
+if [ ! $debug ] ; then
+    echo "Cleaning temporary directory $tmpdir" >> $logfile 2>&1
+    rm $tmpdir/* 2> /dev/null
+
+    if [ $tmpdir_created -eq 1 ]; then
+        echo "Removing temporary directory $tmpdir" >> $logfile 2>&1
+        rmdir $tmpdir 2> /dev/null
+        if [ $? != 0 ]; then
+            echo "Warning: the temporary directory could not be removed." >> $logfile 2>&1
+        fi
     fi
 fi
